# turbotab/server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TurboTab",
    "description": "TurboTab Next: the HTTP and SSE contract between the engine and the frontend.",
    "version": "2.0.0.dev0"
  },
  "paths": {
...
    "/api/fs/list": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "Fs List",
        "description": "The entries of one folder on this machine (local mode only). Hidden entries are left out.",
        "operationId": "fs_list_api_fs_list_get",
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "A folder; default: your home folder",
              "title": "Path"
            },
            "description": "A folder; default: your home folder"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FsListing"
                }
              }
            }
          },
          "403": {
            "description": "Server mode, or a folder TurboTab may not read",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Refusal"
                }
              }
            }
          },
          "404": {
            "description": "No such folder",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Refusal"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Exit": {
        "properties": {
          "label": {
            "type": "string",
            "title": "Label"
          },
          "decision": {
            "anyOf": [
              {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/SetLens"
                  },
                  {
                    "$ref": "#/components/schemas/SetTarget"
                  },
                  {
                    "$ref": "#/components/schemas/SetTask"
                  },
                  {
                    "$ref": "#/components/schemas/SetPurpose"
                  },
                  {
                    "$ref": "#/components/schemas/Revert"
                  }
                ],
                "discriminator": {
                  "propertyName": "kind",
                  "mapping": {
                    "revert": "#/components/schemas/Revert",
                    "set_lens": "#/components/schemas/SetLens",
                    "set_purpose": "#/components/schemas/SetPurpose",
                    "set_target": "#/components/schemas/SetTarget",
                    "set_task": "#/components/schemas/SetTask"
                  }
                }
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision"
          }
        },
        "type": "object",
        "required": [
          "label",
          "decision"
        ],
        "title": "Exit"
      },
      "FsEntry": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "path": {
            "type": "string",
            "title": "Path"
          },
          "is_dir": {
            "type": "boolean",
            "title": "Is Dir"
          },
          "size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size"
          }
        },
        "type": "object",
        "required": [
          "name",
          "path",
          "is_dir",
          "size"
        ],
        "title": "FsEntry"
      },
      "FsListing": {
        "properties": {
          "path": {
            "type": "string",
            "title": "Path"
          },
          "parent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/FsEntry"
            },
            "type": "array",
            "title": "Entries"
          }
        },
        "type": "object",
        "required": [
          "path",
          "parent",
          "entries"
        ],
        "title": "FsListing"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Refusal": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/RefusalDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "Refusal",
        "description": "Every error the API answers itself; HTTP 409 when a decision is refused."
      },
      "RefusalDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "exits": {
            "items": {
              "$ref": "#/components/schemas/Exit"
            },
            "type": "array",
            "title": "Exits"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message",
          "exits"
        ],
        "title": "RefusalDetail"
      },
      "Revert": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "revert",
            "title": "Kind",
            "default": "revert"
          },
          "decision_id": {
            "type": "string",
            "minLength": 1,
            "title": "Decision Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kind",
          "decision_id"
        ],
        "title": "Revert"
      },
      "SetLens": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "set_lens",
            "title": "Kind",
            "default": "set_lens"
          },
          "lenses": {
            "items": {
              "type": "string",
              "enum": [
                "metabolomics",
                "genomics",
                "dietary",
                "clinical",
                "survey"
              ]
            },
            "type": "array",
            "minItems": 1,
            "title": "Lenses"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kind",
          "lenses"
        ],
        "title": "SetLens"
      },
      "SetPurpose": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "set_purpose",
            "title": "Kind",
            "default": "set_purpose"
          },
          "purpose": {
            "type": "string",
            "enum": [
              "prediction",
              "inference"
            ],
            "title": "Purpose"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kind",
          "purpose"
        ],
        "title": "SetPurpose"
      },
      "SetTarget": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "set_target",
            "title": "Kind",
            "default": "set_target"
          },
          "column": {
            "type": "string",
            "minLength": 1,
            "title": "Column"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kind",
          "column"
        ],
        "title": "SetTarget"
      },
      "SetTask": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "set_task",
            "title": "Kind",
            "default": "set_task"
          },
          "task": {
            "type": "string",
            "enum": [
              "regression",
              "binary",
              "multiclass"
            ],
            "title": "Task"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kind",
          "task"
        ],
        "title": "SetTask",
        "description": "An override of task detection."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}